+++ src/cmd/groupby.rs
@@ -3,7 +3,9 @@ use std::num::NonZeroUsize;
 
 use crate::cmd::parallel::Args as ParallelArgs;
 use crate::config::{Config, Delimiter};
-use crate::moonblade::{AggregationProgram, GroupAggregationProgram, GroupPivotAggregationProgram};
+use crate::moonblade::{
+    AggregationProgram, GroupAggregationProgram, GroupBroadcastAggregationProgram,
+};
 use crate::select::SelectColumns;
 use crate::util;
 use crate::CliResult;
@@ -51,10 +53,9 @@ You can group on multiple columns (read `xan select -h` for more information abo
 
 ---
 
-This command is also able to compute a single aggregation over \"pivoted\" columns
-that can be selected through the --pivot <cols> flag. In which case, the aggregation
-clause will recognize a single variable named \"cell\", representing the value contained
-in currently processed column.
+This command is also able to aggregate along columns that you can select using
+the --along-cols <cols> flag. In which case, the aggregation functions will accept
+the anonymous `_` placeholder representing currently processed column's value.
 
 For instance, given the following file:
 
@@ -66,7 +67,7 @@ john,4,6
 
 Using the following command:
 
-    $ xan groupby user --pivot count1,count2 'sum(cell)' file.csv
+    $ xan groupby user --along-cols count1,count2 'sum(cell)' file.csv
 
 Will produce the following result:
 
@@ -87,7 +88,7 @@ For a list of available functions, use `xan help functions`.
 Aggregations can be computed in parallel using the -p/--parallel or -t/--threads flags.
 But this cannot work on streams or gzipped files, unless a `.gzi` index (as created
 by `bgzip -i`) can be found beside it. Parallelization is not compatible
-with the -S/--sorted nor --pivot flags.
+with the -S/--sorted nor -C/--along-cols flags.
 
 Usage:
     xan groupby [options] <column> <expression> [<input>]
@@ -97,10 +98,8 @@ groupby options:
     --keep <cols>            Keep this selection of columns, in addition to
                              the ones representing groups, in the output. Only
                              values from the first seen row per group will be kept.
-    --pivot <cols>           Perform a single aggregation over all of selected columns
-                             where current column value will be named \"cell\" in given
-                             expression and create a column per group with the result in
-                             the output.
+    -C, --along-cols <cols>  Perform a single aggregation over all of selected columns
+                             and create a column per group with the result in the output.
     -S, --sorted             Use this flag to indicate that the file is already sorted on the
                              group columns, in which case the command will be able to considerably
                              optimize memory usage.
@@ -129,7 +128,7 @@ struct Args {
     flag_output: Option<String>,
     flag_delimiter: Option<Delimiter>,
     flag_keep: Option<SelectColumns>,
-    flag_pivot: Option<SelectColumns>,
+    flag_along_cols: Option<SelectColumns>,
     flag_sorted: bool,
     flag_parallel: bool,
     flag_threads: Option<NonZeroUsize>,
@@ -139,8 +138,8 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
     let mut args: Args = util::get_args(USAGE, argv)?;
 
     if args.flag_parallel || args.flag_threads.is_some() {
-        if args.flag_pivot.is_some() {
-            Err("-p/--parallel or -t/--threads cannot be used with --pivot!")?;
+        if args.flag_along_cols.is_some() {
+            Err("-p/--parallel or -t/--threads cannot be used with --along-cols!")?;
         }
 
         if args.flag_sorted {
@@ -171,17 +170,20 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
 
     let sel = rconf.selection(headers)?;
 
-    // --pivot
-    if let Some(selection) = args.flag_pivot.take() {
+    // --along-cols
+    if let Some(selection) = args.flag_along_cols.take() {
         if args.flag_sorted || args.flag_keep.is_some() {
-            Err("--pivot does not work with -S/--sorted nor --keep!")?;
+            Err("--along-cols does not work with -S/--sorted nor --keep!")?;
         }
 
         let mut pivot_sel = selection.selection(headers, !args.flag_no_headers)?;
         pivot_sel.sort_and_dedup();
 
-        let mut program =
-            GroupPivotAggregationProgram::parse(&args.arg_expression, pivot_sel.len())?;
+        let mut program = GroupBroadcastAggregationProgram::parse(
+            &args.arg_expression,
+            headers,
+            pivot_sel.len(),
+        )?;
 
         if !args.flag_no_headers {
             let mut output_headers = sel.select(headers).collect::<csv::ByteRecord>();
@@ -199,7 +201,7 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
         while rdr.read_byte_record(&mut record)? {
             let group = sel.collect(&record);
 
-            program.run_with_cells(group, index, pivot_sel.select(&record))?;
+            program.run_with_cells(group, index, &record, pivot_sel.select(&record))?;
 
             index += 1;
         }